[dace] Address review comments
