+++ src/gt4py/next/otf/languages.py
@@ -57,7 +57,7 @@ class Python(LanguageTag):
     ...
 
 
-class JSON(LanguageTag):
+class SDFG(LanguageTag):
     settings_class = LanguageSettings
     ...
