+++ src/gt4py/next/program_processors/runners/dace_iterator/workflow.py
@@ -15,31 +15,33 @@
 from __future__ import annotations
 
 import dataclasses
-from typing import Callable, Optional
+from typing import Callable, Optional, cast
 
 import dace
+import factory
+from dace.codegen.compiled_sdfg import CompiledSDFG
 
 from gt4py._core import definitions as core_defs
-from gt4py.next import config
+from gt4py.next import common, config
 from gt4py.next.common import Dimension
 from gt4py.next.iterator import ir as itir
 from gt4py.next.iterator.transforms import LiftMode
 from gt4py.next.otf import languages, stages, step_types, workflow
 from gt4py.next.otf.binding import interface
 from gt4py.next.otf.compilation import cache
-from gt4py.next.otf.compilation.compiler import LanguageSettingsType, SourceLanguageType
 from gt4py.next.otf.languages import LanguageSettings
-from gt4py.next.program_processors.runners.dace_iterator import build_sdfg_from_itir
 from gt4py.next.type_system import type_translation as tt
+
+from . import build_sdfg_from_itir, get_sdfg_args
 
 
 @dataclasses.dataclass(frozen=True)
 class DaCeTranslator(
     workflow.ChainableWorkflowMixin[
         stages.ProgramCall,
-        stages.ProgramSource[languages.JSON, languages.LanguageSettings],
+        stages.ProgramSource[languages.SDFG, languages.LanguageSettings],
     ],
-    step_types.TranslationStep[languages.JSON, languages.LanguageSettings],
+    step_types.TranslationStep[languages.SDFG, languages.LanguageSettings],
 ):
     auto_optimize: bool = False
     lift_mode: LiftMode = LiftMode.FORCE_INLINE
@@ -59,8 +61,8 @@ class DaCeTranslator(
     def __call__(
         self,
         inp: stages.ProgramCall,
-    ) -> stages.ProgramSource[languages.JSON, LanguageSettings]:
-        """Generate DaCe SDFG (JSON file) from the ITIR definition."""
+    ) -> stages.ProgramSource[languages.SDFG, LanguageSettings]:
+        """Generate DaCe SDFG file from the ITIR definition."""
         program: itir.FencilDefinition = inp.program
         on_gpu = True if self.device_type == core_defs.DeviceType.CUDA else False
 
@@ -86,29 +88,34 @@ class DaCeTranslator(
             for param, arg in zip(sdfg.arg_names, inp.args)
         )
 
-        module: stages.ProgramSource[languages.JSON, languages.LanguageSettings] = (
+        module: stages.ProgramSource[languages.SDFG, languages.LanguageSettings] = (
             stages.ProgramSource(
                 entry_point=interface.Function(program.id, arg_types),
                 source_code=sdfg.to_json(),
                 library_deps=tuple(),
-                language=languages.JSON,
+                language=languages.SDFG,
                 language_settings=self._language_settings(),
             )
         )
         return module
 
 
+class DaCeTranslationStepFactory(factory.Factory):
+    class Meta:
+        model = DaCeTranslator
+
+
 @dataclasses.dataclass(frozen=True)
 class DaCeCompiler(
     workflow.ChainableWorkflowMixin[
-        stages.CompilableSource[SourceLanguageType, LanguageSettingsType, languages.Python],
+        stages.CompilableSource[languages.SDFG, languages.LanguageSettings, languages.Python],
         stages.CompiledProgram,
     ],
     workflow.ReplaceEnabledWorkflowMixin[
-        stages.CompilableSource[SourceLanguageType, LanguageSettingsType, languages.Python],
+        stages.CompilableSource[languages.SDFG, languages.LanguageSettings, languages.Python],
         stages.CompiledProgram,
     ],
-    step_types.CompilationStep[SourceLanguageType, LanguageSettingsType, languages.Python],
+    step_types.CompilationStep[languages.SDFG, languages.LanguageSettings, languages.Python],
 ):
     """Use the dace build system to compile a GT4Py program to a ``gt4py.next.otf.stages.CompiledProgram``."""
 
@@ -118,7 +125,7 @@ class DaCeCompiler(
 
     def __call__(
         self,
-        inp: stages.CompilableSource[SourceLanguageType, LanguageSettingsType, languages.Python],
+        inp: stages.CompilableSource[languages.SDFG, languages.LanguageSettings, languages.Python],
     ) -> stages.CompiledProgram:
         sdfg = dace.SDFG.from_json(inp.program_source.source_code)
 
@@ -139,3 +146,36 @@ class DaCeCompiler(
             sdfg_program = sdfg.compile(validate=False)
 
         return sdfg_program
+
+
+class DaCeCompilationStepFactory(factory.Factory):
+    class Meta:
+        model = DaCeCompiler
+
+
+def convert_args(
+    inp: stages.CompiledProgram,
+    device: core_defs.DeviceType = core_defs.DeviceType.CPU,
+    use_field_canonical_representation: bool = False,
+) -> stages.CompiledProgram:
+    sdfg_program = cast(CompiledSDFG, inp)
+    on_gpu = True if device == core_defs.DeviceType.CUDA else False
+    sdfg = sdfg_program.sdfg
+
+    def decorated_program(
+        *args, offset_provider: dict[str, common.Connectivity | common.Dimension]
+    ):
+        sdfg_args = get_sdfg_args(
+            sdfg,
+            *args,
+            check_args=False,
+            offset_provider=offset_provider,
+            on_gpu=on_gpu,
+            use_field_canonical_representation=use_field_canonical_representation,
+        )
+
+        with dace.config.temporary_config():
+            dace.config.Config.set("compiler", "allow_view_arguments", value=True)
+            return inp(**sdfg_args)
+
+    return decorated_program
